+++ internal/daemon/hub/src/server/app/app.go
@@ -10,7 +10,6 @@ import (
 	"go.yorun.ai/vine/internal/core/meta"
 	"go.yorun.ai/vine/internal/core/mtls"
 	"go.yorun.ai/vine/internal/daemon"
-	hubapp "go.yorun.ai/vine/internal/daemon/hub/api/app"
 	"go.yorun.ai/vine/internal/daemon/hub/src/server/comp/configaccess"
 	"go.yorun.ai/vine/internal/daemon/hub/src/server/comp/lockserver"
 	"go.yorun.ai/vine/internal/daemon/hub/src/server/comp/natsserver"
@@ -67,7 +66,6 @@ func (a *HubApp) DIInit() {
 		DisableHTTPServer: true,
 		ProtectHTTPServer: true,
 		HTTPServerClients: []mtls.SPIFFEPath{daemon.PortalIdentity.SPIFFEPath()},
-		InprocHostPath:    hubapp.HubAdminInprocHostPath,
 	}
 
 	a.AppFlag.ListenAddr = a.Flag.AdminListen